pyproject.toml
[tool.poetry]
name = "repl_python3_nikityprime"
version = "0.1.0"
description = ""
authors = ["saminsakur <<>>"]

[tool.poetry.dependencies]
python = "^3.8"
discord = "^1.7.3"
python-decouple = "^3.4"
virtualenv = "^20.4.7"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
